Create plotMouseBehaviorAsRaster.m
This MatLab program takes in data that is output from the mice_notes.py program (in seconds), strips the extra characters, and plots in whole seconds the data in a chronological order with color associated with each behavior in a raster style plot. If a behavior does not exist in the data output by the mice_notes.py program, simply comment out the data input, label, and figure sections for that behavior (e.g. no nesting, use % for each line that includes nesting).